Rough implementation of Strafing and attacking
Strafing is determined by the PlayerAnimate class. It reads if the player has "LockedOn" to a target and will determine the animation values based on plaer movement and the direction to the target,

Player attacking is a small lunge towards it's target. The motion is driven by a playerController state. Player lunges in it's heading direction unless it's locked on and will lunge towards the target.