Attack Actions improved. Minor Camera Fix
Attacks are now created and set as Scriptable objects. They work specifically as an IPlayerMoveAction. Which is used by the player controller when an Action should override player movement and inputs. In this example case, the player avatar will Lurch towards it's locked on target or simply lurch forwards without a target.

The camera is now looking at the reticle position of the lockon target rather than the gamelogic position. Ensuring that visual logic is not directly tied into the gameplay logic.