Set up player prefab
Changed and edited many player related scripts so that they are less dependent on outside resources.

They player prefab only needs to connect to a cinemachine brain and the player input component needs to reference the player camera.

A GamePackage prefab also exists where the user can drag and drop into any scene and the neccesary components dor the game to run will be connected to each other.