Enemy Animation, Player Control Update
Enemies and the Player are controlled in a new method by being an IEntity and IActionable.

By considering the player and any enemy as an entity they can share control methods and actions. In this specific case, the player and enemies will be able to peform attacks by making use of the ActionController class to perform an attack which implements the IEntityAction interface.