Weapon Collision Detection
Weapons are a collider that when enabled will begin to detect unique collisions until disabled.

The intention is that a collider will pass through multiple entities and only register a single hit.
The weapon collider specifically searches for a weapon hit receiver attached to a collider. This is used as a sort of "Relay" point to then tell it's referenced IEntity the relevant data.