EnemyDirector Acts as a blackboard, Simple audio on enemies.
Enemy Director acts as a blackboard for enemies. It holds an object pool of enemies and can also used as a means for spawning more.

The main thing it does at the moment is control how often enemies are allowed to attack as a group. This is so the player won't feel overwhelmed.

Simple audio is added to enemies. They play a sound when hit.
Wah.