Payload Motor Moves
Payload moves cohesively.
No more out of range error when using unlooped line.

Payload can die. And explode with rigidbodies.

Entities have a GetAttackPosition which is used to flexibly get the location that AI enemies will use to travel to. Not every entity will have the same boundary. Player and enmies can use a simple circular range. That should be fine for humaoids. but the payload has a capsule instead.