Check point Events For payload
Events are called when the payload moves past a specified distance value.

The value ranges from 0 to 1, and is essentially how far along the spline path where it should occur.

Checkpoints happen sequentially and the list is ordered on awake. I would prefer to make the list ordered when using the inspector in Unity, but that will need to be added a later time.