Gameplay Test Scene + other changes: Action, EntityAnim, Spawner
Gameplay::
Game Manager now has events for when actions occur through the game lifetime.
Currently. StartGame, EndGame, WinGame, LoseGame, and Reset Game.
In This test scene. The player must protect the payload from constantly spawning enemies. When the payload gets to the end, the chest will open and the player "Wins".
The player loses if either the payload dies or the player dies.

Action::
Actions now call Endaction if an action is completed without being cancelled. This is in case additional variables need to be processed. In the test scene SpawnAction needs to reset the anim transform on End.

Animation::
Climb animation used for enemy spawning.
This is an Action that enemies use to spawn from the ground. There's no reason that the player can't use it either, however, the player hasn't implemented it yet.

Spawner::
Enemies had a slight change about how they spawn. The spawn Function in EnemyDirector now requires a position and a heading. This is currently generated randomly by the Enemy spawn box.