Game Package + PlayerHUD
The Game Package scriptable object can be used to create a scene where Game Objects are pre linked together for the game to run.
It would probably be a better idea to create some sort of prefab that creates the necessary objects but this should be fine for now. Also arguably gives a bit more control.

Player Hud is separated from gameManager so that base systems are not interfered with when a HUD is not present.