Quick target lock on
trying to attack a not locked on target was difficult. So I've added a "quick target". the lock on controller will constantly look for a lock on target. However, there is nothing preventing the quick target from switching until the player "Locks on". This will mean the player will always move towards an enemy when attacking. The quick target is currently a yellow reticle and Locked on is red. Variables need to be introduced that determine when a quick target should be able to switch to a new target. Something like a bigger angle threshold. Target finding also needs to be tweaked slighty. Currently the targetfind algorithm searches for the closest enemy in the centre of the player cam.